Move decorators to separate package
