+++ src/main/java/pl/put/poznan/transformer/logic/JsonTransformer.java
@@ -1,0 +1,2 @@
+package pl.put.poznan.transformer.logic;public class JsonTransformer {
+}
