+++ src/main/java/pl/put/poznan/transformer/logic/decorators/JsonClarifierDecorator.java
@@ -1,8 +1,9 @@
-package pl.put.poznan.transformer.logic;
+package pl.put.poznan.transformer.logic.decorators;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import pl.put.poznan.transformer.logic.Json;
 
 public class JsonClarifierDecorator extends JsonDecorator {
     public JsonClarifierDecorator(Json content) {
