+++ src/main/java/pl/put/poznan/transformer/logic/decorators/JsonValidatorDecorator.java
@@ -1,12 +1,12 @@
-package pl.put.poznan.transformer.logic;
+package pl.put.poznan.transformer.logic.decorators;
 
 import java.io.IOException;
-import java.util.LinkedList;
-import java.util.Queue;
 import java.util.Stack;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
+import pl.put.poznan.transformer.logic.Json;
+import pl.put.poznan.transformer.logic.decorators.JsonDecorator;
 
 public class JsonValidatorDecorator extends JsonDecorator {
     public JsonValidatorDecorator(Json content) {
